Added ability to control the capture component using custom states (complance, uploads, sights and settings) and accessing all capture api methods using a ref
