+++ .gitignore
@@ -77,3 +77,6 @@ buck-out/
 
 # @end expo-cli
 demo.dev.monk.yml
+
+# vscode
+.vscode
